Damage Recorder Fallback, With real Damage Implement instead of placeholder
