+++ src/main/java/sypztep/dominatus/common/api/entity/DominatusLivingEntityEvents.java
@@ -22,6 +22,14 @@ public final class DominatusLivingEntityEvents {
         }
         return amount;
     });
+    public static final Event<DamageDealt> DAMAGE_DEALT = EventFactory.createArrayBacked(
+            DamageDealt.class,
+            (listeners) -> (entity, source, finalDamage) -> {
+                for (DamageDealt listener : listeners) {
+                    listener.onDamageDealt(entity, source, finalDamage);
+                }
+            }
+    );
     @FunctionalInterface
     public interface PreArmorDamage {
         /**
@@ -46,5 +54,17 @@ public final class DominatusLivingEntityEvents {
          * @return The modified damage amount (return 0 or negative to cancel damage)
          */
         float modifyDamage(LivingEntity entity, DamageSource source, float amount);
+    }
+    @FunctionalInterface
+    public interface DamageDealt {
+        /**
+         * Called when damage is being dealt to an entity.
+         * This is called after all damage calculations are complete.
+         *
+         * @param entity The entity taking damage
+         * @param source The damage source
+         * @param finalDamage The final damage amount that will be applied
+         */
+        void onDamageDealt(LivingEntity entity, DamageSource source, float finalDamage);
     }
 }
